--- src/__init__.py
@@ -1,6 +1,0 @@
-"""
-Source code for the project.
-"""
-from root_config import *
-from src.server.states import *
-from src.server.models import *
