+++ src/server/__init__.py
@@ -1,6 +1,11 @@
 """
 Server package
 """
+
+__name__ = "server"
+
+__export__ = ["models", "state_machine", "arena_manager", "manager_interface"]
+from .arena_manager import ArenaManager
 from .manager_interface import IManager
-from .arenamanager import ArenaManager
-from .states import StateMachine, StateEnum, BaseState
+from .models import Player, Direction, Base
+from .state_machine import StateMachine, StateMachineConfig
