+++ src/server/state_machine/states/wait_game_start.py
@@ -5,8 +5,8 @@ Handles waiting for players to connect to the arena
 
 from __future__ import annotations
 
-from src.server.states import WaitPlayers
-from src.server.states.possible_states import StateEnum
+from .possible_states import StateEnum
+from .wait_players import WaitPlayers
 
 
 class WaitGameStart(WaitPlayers):
