--- src/server/util/__init__.py
@@ -1,4 +1,0 @@
-"""
-Util package for server side.
-Contains all the utilities used in the server side.
-"""
